Hotfix for building docs (#294)
The `publishGuide` task isn't configured properly, in the sense that it doesn't declare
all its inputs. Recent versions of Gradle are able to capture those problems and report
them as errors: they basically indicate that the task would have accidentally worked in
the past.

This commit just patches the build to introduce proper ordering and create the expected
directories before the task is executed.