dokkaHtmlMultiModule.outputDirectory should be a Directory not File (#520)
Another bite of the cherry fixing https://github.com/micronaut-projects/micronaut-kotlin/issues/507

This seems to be a breaking change in dokka.  It used to accept a file provider, but now
requires a directory provider.  The fix here is to stop mapping the directory to a file
and just use the directory.